+++ CHANGELOG.md
@@ -9,6 +9,11 @@ Reference
 -   `Fixed`  for any bug fixes.
 -   `Security`  in case of vulnerabilities.
 ------ 
+### Version 0.2.0 (_2018-08-01_)
+##### Added
+- Async version for disconnect
+##### Fixed
+- `onConnect()` callback not being fired when calling `connect()`
 ### Version 0.1.0 (_2018-07-29_)
 ##### 
 - Initial source commit